move typeof to own file
