+++ value/index.ts
@@ -1,4 +1,7 @@
-export { Nil, nil, Value } from './value';
-export { Type } from './types';
-export { cast, castType } from './cast';
-export { into, intoType } from './convert';
+import { Nil, nil, Value } from './value';
+import { Type } from './types';
+import { typeOf } from './typeof';
+import { cast, castType } from './cast';
+import { into, intoType } from './convert';
+
+export { Nil, nil, Value, Type, typeOf, cast, castType, into, intoType };
