+++ value/typeof.ts
@@ -1,0 +1,26 @@
+import { BaseException } from '../exceptions';
+import { Type } from './types';
+import { Value, Nil } from './value';
+
+export function typeOf(value: Value): Type {
+  const type = typeof value;
+  if (type === 'boolean') {
+    return Type.Boolean;
+  }
+  if (type === 'number') {
+    if (Number.isInteger(value)) {
+      return Type.Integer;
+    }
+    return Type.Real;
+  }
+  if (type === 'string') {
+    return Type.String;
+  }
+  if (value instanceof Nil) {
+    return Type.Nil;
+  }
+  if (value instanceof BaseException) {
+    return Type.Exception;
+  }
+  return Type.Unknown;
+}
